requirements.txt
Django==4.2.9
djangorestframework==3.14.0
drf-yasg==1.21.7